requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.9.1
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.8.1
certifi==2024.8.30
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap4==2025.6
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.3
distro==1.9.0
Django==5.2.5
django-crispy-forms==2.4
django-plotly-dash==2.5.0
dpd_components==0.2.0
Flask==3.0.3
Flask-Cors==5.0.0
folium==0.20.0
fonttools==4.59.1
geopandas==1.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
kiwisolver==1.4.9
lxml==6.0.0
MarkupSafe==3.0.1
matplotlib==3.10.5
msgpack==1.1.1
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.51.2
packaging==24.2
pandas==2.3.1
pdf2image==1.17.0
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
plotly==6.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.9.2
pydantic_core==2.23.4
pykml==0.2.0
PyMuPDF==1.24.12
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
redis==6.4.0
regex==2024.11.6
requests==2.32.3
retrying==1.4.2
shapely==2.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
sseclient==0.0.27
sseclient-py==1.8.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.4
xyzservices==2025.4.0
zipp==3.23.0
